+++ src/UIComponents/atom/Loading/Loading.js
@@ -1,10 +1,11 @@
 import React from "react";
-import "./Loading.scss";
 import classNames from "classnames";
+import PropTypes from "prop-types";
 
-const Loading = ({ center, type, className, ...others }) => {
+const Loading = ({ center, type, className, skipDefaultClass, ...others }) => {
   const containerClasses = classNames({
-    [className]: true,
+    loading: !skipDefaultClass,
+    [className]: !!className,
     [type]: true,
     "center-container": center,
   });
@@ -13,9 +14,16 @@ const Loading = ({ center, type, className, ...others }) => {
 };
 
 Loading.defaultProps = {
-  className: "loading",
+  className: null,
   center: false,
   type: "default-loader",
+  skipDefaultClass: false,
+};
+
+Loading.propTypes = {
+  center: PropTypes.bool,
+  type: PropTypes.oneOf(["default-loader", "dot-loader"]),
+  skipDefaultClass: PropTypes.bool,
 };
 
 export default Loading;